refactor: refactored dtos package
