+++ src/main/java/com/chat/yourway/controller/ContactController.java
@@ -5,7 +5,7 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 import com.chat.yourway.config.openapi.OpenApiExamples;
 import com.chat.yourway.dto.request.EditContactProfileRequestDto;
-import com.chat.yourway.dto.response.ApiErrorResponseDto;
+import com.chat.yourway.dto.response.error.ApiErrorResponseDto;
 import com.chat.yourway.dto.response.ContactProfileResponseDto;
 import com.chat.yourway.service.interfaces.ContactService;
 import io.swagger.v3.oas.annotations.Operation;
