+++ src/main/java/com/chat/yourway/dto/response/error/ApiErrorResponseDto.java
@@ -1,4 +1,4 @@
-package com.chat.yourway.dto.response;
+package com.chat.yourway.dto.response.error;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
