+++ src/main/java/com/chat/yourway/service/ContactEventServiceImpl.java
@@ -3,7 +3,7 @@ package com.chat.yourway.service;
 import static com.chat.yourway.model.event.EventType.ONLINE;
 import static com.chat.yourway.model.event.EventType.SUBSCRIBED;
 
-import com.chat.yourway.dto.response.LastMessageResponseDto;
+import com.chat.yourway.dto.response.notification.LastMessageResponseDto;
 import com.chat.yourway.model.event.ContactEvent;
 import com.chat.yourway.model.event.EventType;
 import com.chat.yourway.repository.ContactEventRedisRepository;
