+++ src/main/java/com/chat/yourway/service/NotificationServiceImpl.java
@@ -1,7 +1,7 @@
 package com.chat.yourway.service;
 
-import com.chat.yourway.dto.response.MessageNotificationResponseDto;
-import com.chat.yourway.dto.response.TopicNotificationResponseDto;
+import com.chat.yourway.dto.response.notification.MessageNotificationResponseDto;
+import com.chat.yourway.dto.response.notification.TopicNotificationResponseDto;
 import com.chat.yourway.mapper.NotificationMapper;
 import com.chat.yourway.model.event.ContactEvent;
 import com.chat.yourway.service.interfaces.ContactEventService;
