+++ src/test/java/com/chat/yourway/integration/controller/ChatControllerTest.java
@@ -11,8 +11,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.chat.yourway.dto.request.MessagePrivateRequestDto;
 import com.chat.yourway.dto.request.MessagePublicRequestDto;
 import com.chat.yourway.dto.request.PageRequestDto;
-import com.chat.yourway.dto.response.LastMessageResponseDto;
-import com.chat.yourway.dto.response.MessageNotificationResponseDto;
+import com.chat.yourway.dto.response.notification.LastMessageResponseDto;
+import com.chat.yourway.dto.response.notification.MessageNotificationResponseDto;
 import com.chat.yourway.dto.response.MessageResponseDto;
 import com.chat.yourway.integration.controller.websocketclient.TestStompFrameHandler;
 import com.chat.yourway.integration.extension.PostgresExtension;
